[feature/teams]: Updated pyproject.toml file
