+++ pyproject.toml
@@ -7,7 +7,8 @@ fixable = ["ALL"]
 
 [tool.ruff.format]
 skip-magic-trailing-comma = false
-line-ending = "auto"
+line-ending = "lf"
+quote-style = "double"
 
 [tool.black]
 line-length = 120
